pyproject.toml
[tool.poetry]
name = "fruit-ripeness"
version = "0.1.0"
description = "Video Fruit Ripeness Detection App"
authors = ["linhsinp <linhsinp@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.25.1"
scikit-image = "^0.24"
pillow = "^10.0.0"
opencv-python-headless = "^4.10.0.84"
moviepy = "^1.0.3"
ffmpeg-python = "^0.2.0"
inference-sdk = "^0.12.1"
roboflow = "^1.1.32"
supervision = "^0.21.0"
ultralytics = "^8.2.73"
pre-commit = "^4.2.0"
isort = "^6.0.1"
ruff = "^0.11.2"
flask = "^3.1.1"

[build-system]
requires = ["poetry-core=1.8.3"]
build-backend = "poetry.core.masonry.api"
